+++ core/src/cluster_slot_state_verifier.rs
@@ -853,8 +853,11 @@ pub(crate) fn check_slot_agrees_with_cluster(
 mod test {
     use {
         super::*,
-        crate::{progress_map::ProgressMap, replay_stage::tests::setup_forks_from_tree},
+        crate::{replay_stage::tests::setup_forks_from_tree},
         crossbeam_channel::unbounded,
+        solana_consensus::{
+            progress_map::ProgressMap,
+        },
         solana_runtime::bank_forks::BankForks,
         std::{
             collections::{HashMap, HashSet},
