+++ local-cluster/tests/local_cluster_slow_1.rs
@@ -5,11 +5,11 @@ use {
     common::*,
     log::*,
     serial_test::serial,
+    solana_consensus::consensus::SWITCH_FORK_THRESHOLD,
     solana_core::{
         broadcast_stage::{
             broadcast_duplicates_run::BroadcastDuplicatesConfig, BroadcastStageType,
         },
-        consensus::SWITCH_FORK_THRESHOLD,
         replay_stage::DUPLICATE_THRESHOLD,
         validator::ValidatorConfig,
     },
