+++ validator/src/admin_rpc_service.rs
@@ -864,7 +864,7 @@ mod tests {
         rand::{distributions::Uniform, thread_rng, Rng},
         serde_json::Value,
         solana_account_decoder::parse_token::spl_token_pubkey,
-        solana_core::tower_storage::NullTowerStorage,
+        solana_consensus::tower_storage::NullTowerStorage,
         solana_gossip::cluster_info::ClusterInfo,
         solana_ledger::genesis_utils::{create_genesis_config, GenesisConfigInfo},
         solana_rpc::rpc::create_validator_exit,
